Update to imgui 1.54 WIP
From Omar:

The patch in this mail include the fix mentioned in the mail sent
earlier today.

Here are your action:

1) Merge latest master. You *will* get a conflict in this function.
You can solve it with the code below. The part in bold between #if 1 is
the part you can add over the latest origin/master code
(So those bolds line will be part of your patch. You can ditch the #if 1
if you want)

        if (g.IO.MouseDown[0])
        {
            // MovingWindow = window we clicked on, could be a child
window. We track it to preserve Focus and so that ActiveIdWindow ==
MovingWindow and ActiveId == MovingWindow->MoveId for consistency.
            // actually_moving_window = MovingWindow->RootWindow.
            ImGuiWindow* actually_moving_window =
g.MovingWindow->RootWindow;

            // FIXME: Won't work if the tooltip itself has child.
            if ((g.MovingWindow->Flags & ImGuiWindowFlags_ChildWindow)
&& (g.MovingWindow->Flags & ImGuiWindowFlags_Tooltip))
                actually_moving_window = g.MovingWindow;

            ImVec2 pos = g.IO.MousePos - g.ActiveIdClickOffset;
            if (actually_moving_window != g.MovingWindow)
                pos += actually_moving_window->PosFloat -
g.MovingWindow->PosFloat;
            if (actually_moving_window->PosFloat.x != pos.x ||
actually_moving_window->PosFloat.y != pos.y)
            {
                MarkIniSettingsDirty(actually_moving_window);
                actually_moving_window->PosFloat = pos;
            }
            FocusWindow(g.MovingWindow);
        }

2) Then apply attached patch over your "Pinned tooltip" imgui patch

There's are two things in it:
- The child>root connection is cut when a window is both Child and
Tooltip, which fixes the IsItemHovered() issue that you were working
around by clearing MousePos.
- The second fix is the one of my earlier mail which was a conflict
merge issue which broke child windows and the EventList.

3)  A change from a few weeks also created a stronger dependency on the
mouse position.
In TipWindows::update( bool mouse_captured ) you are reseting mouse
position:

                // ...invalidate mouse pos so nobody does stuff under us
                ImGui::GetIO().MousePos = { -FLT_MAX, -FLT_MAX };

It was a workaround against the fact that IsItemHovered() returned true.
With the first line of my patch, this code (and the loop/most of the
function) can now be removed.

Altering MousePos mid-frame is currently problematic because imgui
doesn't latch full event and code in EndFrame() relies on MousePos (it's
a flaw in imgui that i'll aim to correct but probably part of a 2.0)

There's a backward-compatibility issue in IsItemHovered() that's a bit
tricky to fix now (it accepts items from child window), I will
eventually find a nice way to fix it and the first part of my patch
would not be needed anymore.

I will eventually integrate something close to this "child+tooltip"
feature, it's just really error-prone to commit this code as-is in
master (especially with that latest patch).
I have similar issues with some advanced users of Popups where we want
some features of a popup but not others.
I think I may need to separate and isolate features of Child, Tooltip,
and Popups into smaller granularity flags. Common use e.g. _ChildWindow
would pull a few of those behaviors, but we would have finer controls.

(4)
There is an unnecessary #include <future> in imgui.h in your patch,
remnant of an old discarded patch.
If you remove it,
- you need to #include <future> from gpuvis_utils.h,
- and imconfig.h replace #define ImDrawIdx uint32_t by #define ImDrawIdx
unsigned int