Imgui mouse position update (from Omar)
3)  A change from a few weeks also created a stronger dependency on the
mouse position.
In TipWindows::update( bool mouse_captured ) you are reseting mouse
position:

                // ...invalidate mouse pos so nobody does stuff under us
                ImGui::GetIO().MousePos = { -FLT_MAX, -FLT_MAX };

It was a workaround against the fact that IsItemHovered() returned true.
With the first line of my patch, this code (and the loop/most of the
function) can now be removed.

Altering MousePos mid-frame is currently problematic because imgui
doesn't latch full event and code in EndFrame() relies on MousePos (it's
a flaw in imgui that i'll aim to correct but probably part of a 2.0)

There's a backward-compatibility issue in IsItemHovered() that's a bit
tricky to fix now (it accepts items from child window), I will
eventually find a nice way to fix it and the first part of my patch
would not be needed anymore.