Convert the wdat parser to an etl parser
The wdat format was a good point to start a prototype. Since it was
text based it also helped to understand event sources. Not to mention
that working with TDH is horrible in c++.

But now that we have a base of features. Lets cut out the middle man
and begin parsing windows etl files directly. This removes the extra
etl->wdat processing step in the capture tool. It also makes the same
etl files work in gpuvis and gpuview.