trace_utils: Implement GPUVIS_TRACE_BLOCK(F) for C
This commit ports the GPUVIS_TRACE_BLOCK / GPUVIS_TRACE_BLOCKF macros
that already existed for C++ to C (with GNU extensions). That allows
pure C programs on Linux to have a simple way of quickly tracing the
execution duration of a specific scope.

This is how it's implemented:

First, we generalize the C++ class to a struct. In C++, we add the
constructors/destructors to the struct, which makes it functionally the
same as before, but have the ctors/dtors call into _begin/_end
C functions. That way we can share the setup/finalize code between C and
C++ (even though it's very simple).

In C, the same struct is used. For construction, a GNU statement-expression
macro is used to initialize the struct in-line. For destruction, the
`cleanup` GNU attribute is used to execute a function that finalizes the
block.