Present i915-perf data
On top of adding an additional row in the graph view, this also adds a
new panel at the bottom of the window in which the user can look at
the performance counter values associated to a given item of the GPU
timeline.

This code is conditional to compile with the -DUSE_I915_PERF=ON
option.