i915-perf: set a different default colors to GPU timeline events
By default we're currently using the same color for all GPU timeline
events. But some i915 driver versions don't have tracepoint support to
build correlation back to the process that submitted the work, so that
we can use the process' color.

So just assign a default color so that we can at least tell items
apart on the timeline.