Add support for Linux perf JSON export
A patch has been added to patches/ to add JSON export support to perf.
The resulting JSON file can then be loaded into gpuvis on any platform.
Samples are displayed as zero-duration events in the gpuvis timeline.

This introduces a dependency on RapidJSON fetched at build time.
Currently only the Makefile buildsystem supports JSON. Other
buildsystems need to be modified to fetch it and pass the appropriate
preprocessor defines.

Signed-off-by: Nicholas Fraser <nfraser@codeweavers.com>