i915: Fetch timestamp from the library API
The location of a timestamp in the OA report is dependent on the
specific metrics format. The value of the timestamp may also need to be
modified based on OA timestamp frequency. To keep this logic out of
GPUvis, use the library API to get the OA timestamp from the record.

Signed-off-by: Umesh Nerlige Ramappa <umesh.nerlige.ramappa@intel.com>