When parsing tokens from event strings, ensure we find the start of the token
When parsing event strings such as `presentId=xxxx`, we don't want to
parse that as `tid=xxxx`, so we ensure that the beginning of the token
that we find is actually the beginning of the token in the string.  That
is, we assert that before the `tid` must be either whitespace, an open
parens, or the beginning of the string itself.